# contracts/openapi-public.yaml
openapi: 3.1.0
info:
  title: SmartDiag504 Platform API
  version: 0.4.0
  description: >-
    Contrato de las superficies públicas, de operación y administración de SmartDiag504.
    ERPNext conserva la autoridad financiera y de inventario; esta API coordina catálogo,
    reservas, chatbot, pedidos web y proyección operacional.
servers:
  - url: https://api.smartdiag504.com
security: []
...
paths:
...
  /api/v1/chat/sessions:
    post:
      tags: [public-chat]
      operationId: createPublicChatSession
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChatSessionCreate'}
      responses:
        '201':
          description: Sesión opaca y temporal creada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChatSession'}
        '403': {description: Chat público deshabilitado.}
  /api/v1/chat/sessions/{session_id}/messages:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
      - {$ref: '#/components/parameters/ChatSessionToken'}
    get:
      tags: [public-chat]
      operationId: listPublicChatMessages
      responses:
        '200':
          description: Historial autorizado de la sesión.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChatHistory'}
        '401': {description: Token de sesión inválido o expirado.}
    post:
      tags: [public-chat]
      operationId: sendPublicChatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChatMessageCreate'}
      responses:
        '201':
          description: Mensaje procesado.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChatReply'}
        '200':
          description: Repetición idempotente del mismo mensaje.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChatReply'}
        '401': {description: Token de sesión inválido o expirado.}
        '429': {description: Límite de mensajes excedido.}
  /api/v1/chat/sessions/{session_id}/close:
    post:
      tags: [public-chat]
      operationId: closePublicChatSession
      parameters:
        - {$ref: '#/components/parameters/SessionId'}
        - {$ref: '#/components/parameters/ChatSessionToken'}
      responses:
        '200': {description: Sesión cerrada y revocada.}
        '401': {description: Token de sesión inválido o expirado.}
...
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
  parameters:
    SessionId: {name: session_id, in: path, required: true, schema: {type: string, format: uuid}}
    ChatSessionToken: {name: X-Chat-Session-Token, in: header, required: true, schema: {type: string, minLength: 32}}
...
  schemas:
...
    ChatSessionCreate:
      type: object
      properties:
        locale: {type: string, default: es-HN}
        page_url: {type: [string, 'null'], format: uri}
        referrer: {type: [string, 'null'], format: uri}
        accepted_privacy: {type: boolean, const: true}
    ChatSession:
      type: object
      required: [session_id, session_token, expires_at, welcome_message]
      properties:
        session_id: {type: string, format: uuid}
        session_token: {type: string}
        expires_at: {type: string, format: date-time}
        welcome_message: {type: string}
        quick_prompts: {type: array, items: {type: string}}
        privacy_notice: {type: string}
    ChatMessageCreate:
      type: object
      required: [message, client_message_id]
      properties:
        message: {type: string, minLength: 1, maxLength: 2000}
        client_message_id: {type: string, minLength: 8, maxLength: 120}
    ChatMessage:
      type: object
      required: [id, role, content, created_at]
      properties:
        id: {type: string, format: uuid}
        role: {enum: [user, assistant, system]}
        content: {type: string}
        created_at: {type: string, format: date-time}
        audit_id: {type: [string, 'null']}
        mode: {type: [string, 'null']}
        suggested_actions: {type: array, items: {type: string}}
    ChatReply:
      type: object
      required: [session_id, user_message, assistant_message, mode, suggested_actions]
      properties:
        session_id: {type: string, format: uuid}
        user_message: {$ref: '#/components/schemas/ChatMessage'}
        assistant_message: {$ref: '#/components/schemas/ChatMessage'}
        audit_id: {type: [string, 'null']}
        mode: {type: string}
        suggested_actions: {type: array, items: {type: string}}
    ChatHistory:
      type: object
      required: [session_id, messages]
      properties:
        session_id: {type: string, format: uuid}
        messages:
          type: array
          items: {$ref: '#/components/schemas/ChatMessage'}
...